more fixes after merge
