+++ auth/api/iam/dpop_test.go
@@ -19,6 +19,7 @@
 package iam
 
 import (
+	"context"
 	crypto2 "crypto"
 	"crypto/ecdsa"
 	"crypto/elliptic"
@@ -29,11 +30,188 @@ import (
 
 	"github.com/lestrrat-go/jwx/v2/jwa"
 	"github.com/lestrrat-go/jwx/v2/jwk"
+	ssi "github.com/nuts-foundation/go-did"
+	"github.com/nuts-foundation/go-did/did"
 	"github.com/nuts-foundation/nuts-node/auth/oauth"
+	cryptoNuts "github.com/nuts-foundation/nuts-node/crypto"
 	"github.com/nuts-foundation/nuts-node/crypto/dpop"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/mock/gomock"
 )
+
+func TestWrapper_CreateDPoPProof(t *testing.T) {
+	accesstoken := "token"
+	request, _ := http.NewRequestWithContext(context.Background(), http.MethodGet, "https://example.com", nil)
+	requestBody := CreateDPoPProofJSONRequestBody{
+		Htm:   "GET",
+		Token: accesstoken,
+		Htu:   "https://example.com",
+	}
+	requestObject := CreateDPoPProofRequestObject{
+		Body: &requestBody,
+		Did:  webDID.String(),
+	}
+	didDocument := did.Document{ID: holderDID}
+	vmId := did.MustParseDIDURL(webDID.String() + "#key1")
+	key := cryptoNuts.NewTestKey(vmId.String())
+	vm, _ := did.NewVerificationMethod(vmId, ssi.JsonWebKey2020, webDID, key.Public())
+	didDocument.AddAssertionMethod(vm)
+	dpopToken := dpop.New(*request)
+	dpopToken.GenerateProof(accesstoken)
+	t.Run("ok", func(t *testing.T) {
+		ctx := newTestClient(t)
+		ctx.vdr.EXPECT().IsOwner(gomock.Any(), webDID).Return(true, nil)
+		ctx.resolver.EXPECT().Resolve(webDID, gomock.Any()).Return(&didDocument, nil, nil)
+		ctx.jwtSigner.EXPECT().SignDPoP(gomock.Any(), gomock.Any(), vmId.String()).DoAndReturn(func(_ context.Context, token dpop.DPoP, _ string) (string, error) {
+			assert.Equal(t, dpopToken.String(), token.String())
+			return "dpop", nil
+		})
+
+		res, err := ctx.client.CreateDPoPProof(context.Background(), requestObject)
+
+		require.NoError(t, err)
+		assert.Equal(t, "dpop", res.(CreateDPoPProof200JSONResponse).Dpop)
+	})
+	t.Run("missing method", func(t *testing.T) {
+		ctx := newTestClient(t)
+		requestBody.Htm = ""
+		defer (func() { requestBody.Htm = "GET" })()
+
+		_, err := ctx.client.CreateDPoPProof(context.Background(), requestObject)
+
+		assert.EqualError(t, err, "missing method")
+	})
+	t.Run("invalid method", func(t *testing.T) {
+		ctx := newTestClient(t)
+		ctx.vdr.EXPECT().IsOwner(gomock.Any(), webDID).Return(true, nil)
+		requestBody.Htm = "\\"
+		defer (func() { requestBody.Htm = "GET" })()
+
+		_, err := ctx.client.CreateDPoPProof(context.Background(), requestObject)
+
+		assert.EqualError(t, err, "net/http: invalid method \"\\\\\"")
+	})
+	t.Run("missing token", func(t *testing.T) {
+		ctx := newTestClient(t)
+		requestBody.Token = ""
+		defer (func() { requestBody.Token = accesstoken })()
+
+		_, err := ctx.client.CreateDPoPProof(context.Background(), requestObject)
+
+		assert.EqualError(t, err, "missing token")
+	})
+	t.Run("missing url", func(t *testing.T) {
+		ctx := newTestClient(t)
+		requestBody.Htu = ""
+		defer (func() { requestBody.Htu = "https://example.com" })()
+
+		_, err := ctx.client.CreateDPoPProof(context.Background(), requestObject)
+
+		assert.EqualError(t, err, "missing url")
+	})
+	t.Run("did not owned", func(t *testing.T) {
+		ctx := newTestClient(t)
+		ctx.vdr.EXPECT().IsOwner(gomock.Any(), webDID).Return(false, nil)
+
+		_, err := ctx.client.CreateDPoPProof(context.Background(), requestObject)
+
+		assert.EqualError(t, err, "DID document not managed by this node")
+	})
+	t.Run("proof error", func(t *testing.T) {
+		ctx := newTestClient(t)
+		ctx.vdr.EXPECT().IsOwner(gomock.Any(), webDID).Return(true, nil)
+		ctx.resolver.EXPECT().Resolve(webDID, gomock.Any()).Return(&didDocument, nil, nil)
+		ctx.jwtSigner.EXPECT().SignDPoP(gomock.Any(), gomock.Any(), vmId.String()).Return("dpop", assert.AnError)
+
+		_, err := ctx.client.CreateDPoPProof(context.Background(), requestObject)
+
+		assert.Equal(t, assert.AnError, err)
+	})
+}
+
+func TestWrapper_ValidateDPoPProof(t *testing.T) {
+	accessToken := "token"
+	dpopToken, dpopProof, thumbprint := newSignedTestDPoP()
+	request := ValidateDPoPProofRequestObject{
+		Body: &ValidateDPoPProofJSONRequestBody{
+			DpopProof:  dpopProof.String(),
+			Method:     "POST",
+			Thumbprint: thumbprint,
+			Token:      accessToken,
+			Url:        "https://server.example.com/token",
+		},
+	}
+
+	t.Run("ok", func(t *testing.T) {
+		ctx := newTestClient(t)
+
+		resp, err := ctx.client.ValidateDPoPProof(nil, request)
+
+		require.NoError(t, err)
+		require.IsType(t, ValidateDPoPProof200JSONResponse{}, resp)
+		assert.True(t, resp.(ValidateDPoPProof200JSONResponse).Valid)
+	})
+	t.Run("no match", func(t *testing.T) {
+		ctx := newTestClient(t)
+		request.Body.Method = "GET"
+		defer (func() { request.Body.Method = "POST" })()
+
+		resp, err := ctx.client.ValidateDPoPProof(nil, request)
+
+		require.NoError(t, err)
+		require.IsType(t, ValidateDPoPProof200JSONResponse{}, resp)
+		assert.False(t, resp.(ValidateDPoPProof200JSONResponse).Valid)
+		assert.Equal(t, "method mismatch, token: POST, given: GET", *resp.(ValidateDPoPProof200JSONResponse).Reason)
+	})
+	t.Run("missing ath header", func(t *testing.T) {
+		ctx := newTestClient(t)
+		request.Body.DpopProof = dpopToken.String()
+		defer (func() { request.Body.DpopProof = dpopProof.String() })()
+
+		resp, err := ctx.client.ValidateDPoPProof(nil, request)
+
+		require.NoError(t, err)
+		require.IsType(t, ValidateDPoPProof200JSONResponse{}, resp)
+		assert.False(t, resp.(ValidateDPoPProof200JSONResponse).Valid)
+		assert.Equal(t, "missing ath claim", *resp.(ValidateDPoPProof200JSONResponse).Reason)
+	})
+	t.Run("parsing failed", func(t *testing.T) {
+		ctx := newTestClient(t)
+		request.Body.DpopProof = "invalid"
+		defer (func() { request.Body.DpopProof = dpopProof.String() })()
+
+		resp, err := ctx.client.ValidateDPoPProof(nil, request)
+
+		require.NoError(t, err)
+		require.IsType(t, ValidateDPoPProof200JSONResponse{}, resp)
+		assert.False(t, resp.(ValidateDPoPProof200JSONResponse).Valid)
+		assert.Equal(t, "failed to parse DPoP header: invalid DPoP token\ninvalid compact serialization format: invalid number of segments", *resp.(ValidateDPoPProof200JSONResponse).Reason)
+	})
+	t.Run("invalid accestoken", func(t *testing.T) {
+		ctx := newTestClient(t)
+		request.Body.Token = "invalid"
+		defer (func() { request.Body.Token = accessToken })()
+
+		resp, err := ctx.client.ValidateDPoPProof(nil, request)
+
+		require.NoError(t, err)
+		require.IsType(t, ValidateDPoPProof200JSONResponse{}, resp)
+		assert.False(t, resp.(ValidateDPoPProof200JSONResponse).Valid)
+		assert.Equal(t, "ath/token claim mismatch", *resp.(ValidateDPoPProof200JSONResponse).Reason)
+	})
+	t.Run("already used once", func(t *testing.T) {
+		ctx := newTestClient(t)
+		_ = ctx.client.useNonceOnceStore().Put(dpopProof.Token.JwtID(), struct{}{})
+
+		resp, err := ctx.client.ValidateDPoPProof(nil, request)
+
+		require.NoError(t, err)
+		require.IsType(t, ValidateDPoPProof200JSONResponse{}, resp)
+		assert.False(t, resp.(ValidateDPoPProof200JSONResponse).Valid)
+		assert.Equal(t, "jti already used", *resp.(ValidateDPoPProof200JSONResponse).Reason)
+	})
+}
 
 func Test_dpopFromRequest(t *testing.T) {
 	t.Run("without DPoP header", func(t *testing.T) {
